requirements.txt
pyfst==0.2.3
pyltp==0.1.9.1
argparse==1.4.0
gpustat==0.4.1
torch==0.3.1.post2
torchvision==0.2.0